🚑 sql data type

+++ src/infra/repositories/entry_repository.rs
@@ -156,7 +156,7 @@ pub async fn get_median_entries(
         SELECT
             source,
             MAX(timestamp) AS time,
-            PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY price) AS median_price
+            (PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY price))::numeric AS median_price
         FROM
             entries
         WHERE
